# Reorder span iterators